extracted map from cyphermap
