+++ src/Types/CypherMap.php
@@ -13,444 +13,21 @@ declare(strict_types=1);
 
 namespace Laudis\Neo4j\Types;
 
-use function array_key_exists;
-use function array_key_first;
-use function array_key_last;
-use function array_keys;
-use function array_reverse;
-use function array_slice;
 use function func_num_args;
-use InvalidArgumentException;
-use function is_int;
-use function is_object;
-use function is_string;
-use function ksort;
-use Laudis\Neo4j\Databags\Pair;
 use Laudis\Neo4j\Exception\RuntimeTypeException;
 use Laudis\Neo4j\TypeCaster;
-use function method_exists;
-use OutOfBoundsException;
-use stdClass;
-use function uasort;
-use function uksort;
 
 /**
  * An immutable ordered map of items.
  *
  * @template TValue
  *
- * @extends AbstractCypherSequence<string, TValue>
+ * @extends Map<TValue>
  *
  * @psalm-immutable
  */
-final class CypherMap extends AbstractCypherSequence
+final class CypherMap extends Map
 {
-    /**
-     * @pure
-     */
-    public static function fromIterable(iterable $iterable): AbstractCypherSequence
-    {
-        return new self($iterable);
-    }
-
-    /**
-     * @param iterable<TValue> $iterable
-     */
-    public function __construct(iterable $iterable = [])
-    {
-        if ($iterable instanceof self) {
-            /** @psalm-suppress InvalidPropertyAssignmentValue */
-            $this->sequence = $iterable->sequence;
-        } else {
-            $this->sequence = [];
-            /** @var mixed $key */
-            foreach ($iterable as $key => $value) {
-                if ($key === null || is_int($key) || (is_object($key) && method_exists($key, '__toString'))) {
-                    $this->sequence[(string) $key] = $value;
-                } elseif (is_string($key)) {
-                    $this->sequence[$key] = $value;
-                } else {
-                    throw new InvalidArgumentException('Iterable must have a stringable keys');
-                }
-            }
-        }
-    }
-
-    /**
-     * Returns the first pair in the map.
-     *
-     * @return Pair<string, TValue>
-     */
-    public function first(): Pair
-    {
-        $key = array_key_first($this->sequence);
-        if (!is_string($key)) {
-            throw new OutOfBoundsException('Cannot grab first element of an empty map');
-        }
-
-        return new Pair($key, $this->sequence[$key]);
-    }
-
-    /**
-     * Returns the last pair in the map.
-     *
-     * @return Pair<string, TValue>
-     */
-    public function last(): Pair
-    {
-        $key = array_key_last($this->sequence);
-        if (!is_string($key)) {
-            throw new OutOfBoundsException('Cannot grab last element of an empty map');
-        }
-
-        return new Pair($key, $this->sequence[$key]);
-    }
-
-    /**
-     * Returns the pair at the nth position of the map.
-     *
-     * @return Pair<string, TValue>
-     */
-    public function skip(int $position): Pair
-    {
-        $keys = $this->keys();
-
-        if ($keys->count() > $position) {
-            $key = $keys[$position];
-
-            return new Pair($key, $this->sequence[$key]);
-        }
-
-        throw new OutOfBoundsException(sprintf('Cannot skip to a pair at position: %s', $position));
-    }
-
-    /**
-     * Returns the keys in the map in order.
-     *
-     * @return CypherList<string>
-     */
-    public function keys(): CypherList
-    {
-        return new CypherList(array_keys($this->sequence));
-    }
-
-    /**
-     * Returns the pairs in the map in order.
-     *
-     * @return CypherList<Pair<string, TValue>>
-     */
-    public function pairs(): CypherList
-    {
-        $tbr = [];
-        foreach ($this->sequence as $key => $value) {
-            $tbr[] = new Pair($key, $value);
-        }
-
-        return new CypherList($tbr);
-    }
-
-    /**
-     * Create a new map sorted by keys. Natural ordering will be used if no comparator is provided.
-     *
-     * @param (callable(string, string):int)|null $comparator
-     *
-     * @return CypherMap<TValue>
-     */
-    public function ksorted(callable $comparator = null): CypherMap
-    {
-        $tbr = $this->sequence;
-        if ($comparator === null) {
-            ksort($tbr);
-        } else {
-            /** @psalm-suppress ImpureFunctionCall */
-            uksort($tbr, $comparator);
-        }
-
-        return new self($tbr);
-    }
-
-    /**
-     * Returns the values in the map in order.
-     *
-     * @return CypherList<TValue>
-     */
-    public function values(): CypherList
-    {
-        return new CypherList($this->sequence);
-    }
-
-    /**
-     * Creates a new map using exclusive or on the keys.
-     *
-     * @param iterable<array-key, TValue> $map
-     *
-     * @return CypherMap<TValue>
-     */
-    public function xor(iterable $map): CypherMap
-    {
-        $tbr = $this->sequence;
-
-        foreach ($map as $key => $value) {
-            if (array_key_exists($key, $this->sequence)) {
-                unset($tbr[(string) $key]);
-            } else {
-                $tbr[(string) $key] = $value;
-            }
-        }
-
-        return new self($tbr);
-    }
-
-    /**
-     * @param iterable<array-key, TValue> $values
-     *
-     * @return static
-     */
-    public function merge(iterable $values): self
-    {
-        $other = new self($values);
-        $tbr = $this->sequence;
-
-        foreach ($other as $key => $value) {
-            $tbr[$key] = $value;
-        }
-
-        return new self($tbr);
-    }
-
-    /**
-     * Creates a union of this and the provided map. The items in the original map take precedence.
-     *
-     * @param iterable<array-key, TValue> $map
-     *
-     * @return CypherMap<TValue>
-     */
-    public function union(iterable $map): CypherMap
-    {
-        $tbr = $this->sequence;
-        foreach ($map as $key => $value) {
-            if (!array_key_exists($key, $tbr)) {
-                $tbr[(string) $key] = $value;
-            }
-        }
-
-        return new self($tbr);
-    }
-
-    /**
-     * Creates a new map from the existing one filtering the values based on the keys that don't exist in the provided map.
-     *
-     * @param iterable<array-key, TValue> $map
-     *
-     * @return static
-     */
-    public function intersect(iterable $map): self
-    {
-        $tbr = [];
-        // @psalm-suppress UnusedForeachValue
-        foreach ($map as $key => $value) {
-            if (array_key_exists($key, $this->sequence)) {
-                $tbr[$key] = $this->sequence[$key];
-            }
-        }
-
-        return $this::fromIterable($tbr);
-    }
-
-    /**
-     * Creates a new map from the existing one filtering the values based on the keys that also exist in the provided map.
-     *
-     * @param iterable<array-key, TValue> $map
-     *
-     * @return CypherMap<TValue>
-     */
-    public function diff(iterable $map): CypherMap
-    {
-        $tbr = $this->sequence;
-
-        /** @psalm-suppress UnusedForeachValue */
-        foreach ($map as $key => $value) {
-            unset($tbr[(string) $key]);
-        }
-
-        return new self($tbr);
-    }
-
-    /**
-     * @return static
-     */
-    public function reversed(): self
-    {
-        return $this::fromIterable(array_reverse($this->sequence, true));
-    }
-
-    /**
-     * @return CypherMap<TValue>
-     */
-    public function slice(int $offset, int $length = null): self
-    {
-        return new self(array_slice($this->sequence, $offset, $length, true));
-    }
-
-    /**
-     * @param (pure-callable(TValue, TValue):int)|null $comparator
-     *
-     * @return static
-     */
-    public function sorted(?callable $comparator = null): self
-    {
-        $tbr = $this->sequence;
-        if ($comparator === null) {
-            asort($tbr);
-        } else {
-            uasort($tbr, $comparator);
-        }
-
-        return new self($tbr);
-    }
-
-    /**
-     * Gets the value with the provided key. If a default value is provided, it will return the default instead of throwing an error when the key does not exist.
-     *
-     * @template TDefault
-     *
-     * @param TDefault $default
-     *
-     * @throws OutOfBoundsException
-     *
-     * @return (func_num_args() is 1 ? TValue : TValue|TDefault)
-     */
-    public function get(string $key, $default = null)
-    {
-        if (func_num_args() === 1) {
-            if (!array_key_exists($key, $this->sequence)) {
-                throw new OutOfBoundsException(sprintf('Cannot get item in sequence with key: %s', $key));
-            }
-
-            return $this->sequence[$key];
-        }
-
-        return $this->sequence[$key] ?? $default;
-    }
-
-    public function jsonSerialize()
-    {
-        if ($this->isEmpty()) {
-            return new stdClass();
-        }
-
-        return parent::jsonSerialize();
-    }
-
-    /**
-     * @param mixed $default
-     */
-    public function getAsString(string $key, $default = null): string
-    {
-        if (func_num_args() === 1) {
-            $value = $this->get($key);
-        } else {
-            $value = $this->get($key, $default);
-        }
-        $tbr = TypeCaster::toString($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'string');
-        }
-
-        return $tbr;
-    }
-
-    /**
-     * @param mixed $default
-     */
-    public function getAsInt(string $key, $default): int
-    {
-        if (func_num_args() === 1) {
-            $value = $this->get($key);
-        } else {
-            $value = $this->get($key, $default);
-        }
-        $tbr = TypeCaster::toInt($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'int');
-        }
-
-        return $tbr;
-    }
-
-    /**
-     * @param mixed $default
-     */
-    public function getAsFloat(int $key, $default = null): float
-    {
-        if (func_num_args() === 1) {
-            $value = $this->get($key);
-        } else {
-            $value = $this->get($key, $default);
-        }
-        $tbr = TypeCaster::toFloat($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'float');
-        }
-
-        return $tbr;
-    }
-
-    /**
-     * @param mixed $default
-     */
-    public function getAsBool(int $key, $default = null): bool
-    {
-        if (func_num_args() === 1) {
-            $value = $this->get($key);
-        } else {
-            $value = $this->get($key, $default);
-        }
-        $tbr = TypeCaster::toBool($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'bool');
-        }
-
-        return $tbr;
-    }
-
-    /**
-     * @param mixed $default
-     *
-     * @return null
-     */
-    public function getAsNull(int $key, $default = null)
-    {
-        if (func_num_args() === 1) {
-            $this->get($key);
-        }
-
-        return TypeCaster::toNull();
-    }
-
-    /**
-     * @template U
-     *
-     * @param class-string<U> $class
-     * @param mixed           $default
-     *
-     * @return U
-     */
-    public function getAsObject(int $key, string $class, $default = null): object
-    {
-        if (func_num_args() === 1) {
-            $value = $this->get($key);
-        } else {
-            $value = $this->get($key, $default);
-        }
-        $tbr = TypeCaster::toClass($value, $class);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, $class);
-        }
-
-        return $tbr;
-    }
-
     /**
      * @param mixed $default
      *
@@ -499,6 +76,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, Date::class);
         }
+
         return $this->getAsObject($key, Date::class, $default);
     }
 
@@ -510,6 +88,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, DateTime::class);
         }
+
         return $this->getAsObject($key, DateTime::class, $default);
     }
 
@@ -521,6 +100,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, Duration::class);
         }
+
         return $this->getAsObject($key, Duration::class, $default);
     }
 
@@ -532,6 +112,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, LocalDateTime::class);
         }
+
         return $this->getAsObject($key, LocalDateTime::class, $default);
     }
 
@@ -543,6 +124,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, LocalTime::class);
         }
+
         return $this->getAsObject($key, LocalTime::class, $default);
     }
 
@@ -554,6 +136,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, Time::class);
         }
+
         return $this->getAsObject($key, Time::class, $default);
     }
 
@@ -565,6 +148,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, Node::class);
         }
+
         return $this->getAsObject($key, Node::class, $default);
     }
 
@@ -576,6 +160,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, Relationship::class);
         }
+
         return $this->getAsObject($key, Relationship::class, $default);
     }
 
@@ -587,6 +172,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, Path::class);
         }
+
         return $this->getAsObject($key, Path::class, $default);
     }
 
@@ -598,6 +184,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, Cartesian3DPoint::class);
         }
+
         return $this->getAsObject($key, Cartesian3DPoint::class, $default);
     }
 
@@ -609,6 +196,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, CartesianPoint::class);
         }
+
         return $this->getAsObject($key, CartesianPoint::class, $default);
     }
 
@@ -620,6 +208,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, WGS84Point::class);
         }
+
         return $this->getAsObject($key, WGS84Point::class, $default);
     }
 
@@ -631,6 +220,7 @@ final class CypherMap extends AbstractCypherSequence
         if (func_num_args() === 1) {
             return $this->getAsObject($key, WGS843DPoint::class);
         }
+
         return $this->getAsObject($key, WGS843DPoint::class, $default);
     }
 }
